pyproject.toml
[project]
name = "neurovolume_testing"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11.13" 
dependencies = [
    "bpy>=5.0.1",
    "nibabel>=5.4.2",
    "neurovolume",
    "ruff>=0.15.6",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.6",
]
# probably need to build this before running
# uv sync
# i would guess...
[tool.uv.sources]
neurovolume = { path = "./neurovolume", editable = true}

